ov006: promote dScMgCoin_c segment 2 into a single C++ TU (33 of 61)
Folds ROM ordinals 3..35 of the 0x020dbe40..0x020de940 dScMgCoin_c linker
run -- 0x020dbf7c..0x020dd0e0, 33 one-function shards -- into
src/actors/dScMgCoin_c.cpp. Text-only; verify 33/33 MATCH, objisolate clean,
ROM-ascending emission order; linkcheck LINK-VERIFIED with 106/106 modules
byte-exact, [4b/8] 33 LICENSED / 0 unlicensed / orderOk, and a ROM sha256
identical to the stock control.

Why a sub-range. Two functions inside the run, func_ov006_020dbe9c
(ordinal 2) and func_ov006_020dd0e0 (ordinal 36), have no source and no
delinks.txt entry at all; both are long-standing banked near-misses (7 and
34 divergences), so neither could be brought in. Nothing in this tree can
express a .text claim with a hole -- 0 of 148 TU manifests and 0 delink
blocks carry two .text runs -- so the run splits into 2 / 33 / 26 sections
and only one side can be licensed. This is the largest: 33 beats segment
3's ceiling of 26 even with dScMgCoin_c_classInit folded in. The cost is
that segment 3 keeps four of the six own vtable overrides and the factory.

The queue's pragma:2 blocker is a phantom, measured with the
delete-outright control: the two shards' opt_common_subs /
opt_strength_reduction pragmas are byte-identical to having no pragmas at
all (33/33 both ways). They are gone from the shipped file.
#pragma defer_codegen off stays -- dropping it costs no bytes but inverts
emission order to 32 out-of-order pairs, which linkcheck [4b/8] refuses.

This TU does not own the class's key function (the out-of-line destructor
at 0x020dbe40 is on the far side of the first hole), so it emits no
vtable or RTTI, needs no compiler_only_output rows, and include/dScMgCoin_c.h
is untouched. Ratchet-neutral: none of the 33 shards is in
converted-baseline.json, so tiers_ratchet --update was deliberately not run.
