+++ config/arm9/overlays/ov006/delinks.txt
@@ -1667,137 +1667,9 @@ src/_ZN11dScMgCoin_cD0Ev.cpp:
     complete
     .text start:0x020dbe64 end:0x020dbe9c
 
-src/func_ov006_020dbf7c.c:
+src/actors/dScMgCoin_c.cpp:
     complete
-    .text start:0x020dbf7c end:0x020dc154
-
-src/func_ov006_020dc154.c:
-    complete
-    .text start:0x020dc154 end:0x020dc1c4
-
-src/func_ov006_020dc1c4.c:
-    complete
-    .text start:0x020dc1c4 end:0x020dc26c
-
-src/func_ov006_020dc26c.c:
-    complete
-    .text start:0x020dc26c end:0x020dc294
-
-src/func_ov006_020dc294.c:
-    complete
-    .text start:0x020dc294 end:0x020dc298
-
-src/func_ov006_020dc298.cpp:
-    complete
-    .text start:0x020dc298 end:0x020dc2f8
-
-src/func_ov006_020dc2f8.c:
-    complete
-    .text start:0x020dc2f8 end:0x020dc334
-
-src/func_ov006_020dc334.c:
-    complete
-    .text start:0x020dc334 end:0x020dc348
-
-src/func_ov006_020dc348.c:
-    complete
-    .text start:0x020dc348 end:0x020dc370
-
-src/func_ov006_020dc370.c:
-    complete
-    .text start:0x020dc370 end:0x020dc3bc
-
-src/func_ov006_020dc3bc.c:
-    complete
-    .text start:0x020dc3bc end:0x020dc414
-
-src/func_ov006_020dc414.c:
-    complete
-    .text start:0x020dc414 end:0x020dc4b0
-
-src/func_ov006_020dc4b0.c:
-    complete
-    .text start:0x020dc4b0 end:0x020dc4c8
-
-src/func_ov006_020dc4c8.c:
-    complete
-    .text start:0x020dc4c8 end:0x020dc5c4
-
-src/func_ov006_020dc5c4.c:
-    complete
-    .text start:0x020dc5c4 end:0x020dc6d0
-
-src/func_ov006_020dc6d0.c:
-    complete
-    .text start:0x020dc6d0 end:0x020dc754
-
-src/func_ov006_020dc754.cpp:
-    complete
-    .text start:0x020dc754 end:0x020dc7b4
-
-src/func_ov006_020dc7b4.c:
-    complete
-    .text start:0x020dc7b4 end:0x020dc7fc
-
-src/func_ov006_020dc7fc.c:
-    complete
-    .text start:0x020dc7fc end:0x020dc814
-
-src/func_ov006_020dc814.c:
-    complete
-    .text start:0x020dc814 end:0x020dc870
-
-src/func_ov006_020dc870.c:
-    complete
-    .text start:0x020dc870 end:0x020dc900
-
-src/func_ov006_020dc900.c:
-    complete
-    .text start:0x020dc900 end:0x020dc960
-
-src/func_ov006_020dc960.c:
-    complete
-    .text start:0x020dc960 end:0x020dc99c
-
-src/func_ov006_020dc99c.c:
-    complete
-    .text start:0x020dc99c end:0x020dca04
-
-src/func_ov006_020dca04.c:
-    complete
-    .text start:0x020dca04 end:0x020dcb1c
-
-src/func_ov006_020dcb1c.c:
-    complete
-    .text start:0x020dcb1c end:0x020dcc48
-
-src/func_ov006_020dcc48.c:
-    complete
-    .text start:0x020dcc48 end:0x020dccb8
-
-src/func_ov006_020dccb8.c:
-    complete
-    .text start:0x020dccb8 end:0x020dcd74
-
-src/func_ov006_020dcd74.c:
-    complete
-    .text start:0x020dcd74 end:0x020dce3c
-
-src/func_ov006_020dce3c.c:
-    complete
-    .text start:0x020dce3c end:0x020dcea8
-
-src/func_ov006_020dcea8.c:
-    complete
-    .text start:0x020dcea8 end:0x020dcffc
-
-src/func_ov006_020dcffc.c:
-    complete
-    .text start:0x020dcffc end:0x020dd000
-
-src/func_ov006_020dd000.c:
-    complete
-    .text start:0x020dd000 end:0x020dd0e0
+    .text start:0x020dbf7c end:0x020dd0e0
 
 src/func_ov006_020dd2cc.cpp:
     complete
