--- src/func_ov006_020dc26c.c
@@ -1,6 +1,0 @@
-void func_ov006_020dc26c(char *c){
-  *(char*)(c+0x51bf)=1;
-  *(char*)(c+0x51be)=1;
-  *(int*)(c+0x51b0)=0;
-  *(int*)(c+0x51b4)=-0x4000;
-}
