--- src/func_ov006_020dc2f8.c
@@ -1,9 +1,0 @@
-void func_ov006_020dc2f8(char *c){
-  *(char*)(c+0x51bc)=1;
-  *(char*)(c+0x51bd)=1;
-  *(char*)(c+0x51be)=0;
-  *(char*)(c+0x51bf)=0x43000;
-  *(short*)(c+0x51b8)=0;
-  *(int*)(c+0x51a8)=0x8e000;
-  *(int*)(c+0x51ac)=-0x43000;
-}
