--- src/func_ov006_020dc334.c
@@ -1,5 +1,0 @@
-void func_ov006_020dc334(char *p)
-{
-    *(char *)(p + 0x51bc) = 0;
-    *(char *)(p + 0x51bd) = 0;
-}
