--- src/func_ov006_020dc348.c
@@ -1,6 +1,0 @@
-void func_ov006_020dc348(char *c){
-  *(char*)(c+0x51a1)=4;
-  *(char*)(c+0x51a4)=0;
-  *(short*)(c+0x519e)=0;
-  *(char*)(c+0x51de)=1;
-}
