--- src/func_ov006_020dc4b0.c
@@ -1,3 +1,0 @@
-void func_ov006_020dc4b0(char* c, int i){
-  *(char*)(c + i*0x14 + 0x5000 + 0x1a3) = 0;
-}
