--- src/func_ov006_020dc7b4.c
@@ -1,11 +1,0 @@
-void func_ov006_020dc7b4(char* c){
-  *(char*)(c+0x5000+0x1a0)=1;
-  *(char*)(c+0x5000+0x1a1)=0;
-  *(char*)(c+0x5000+0x1a2)=1;
-  *(short*)(c+0x5000+0x19c)=0x28;
-  *(short*)(c+0x5000+0x19e)=0;
-  *(char*)(c+0x5000+0x1a3)=3;
-  *(char*)(c+0x5000+0x1a4)=0;
-  *(int*)(c+0x5000+0x194)=0x70000;
-  *(int*)(c+0x5000+0x198)=0x98000;
-}
