--- src/func_ov006_020dc7fc.c
@@ -1,5 +1,0 @@
-void func_ov006_020dc7fc(char *p)
-{
-    *(char *)(p + 0x51a0) = 0;
-    *(char *)(p + 0x51a2) = 1;
-}
