--- src/func_ov006_020dcffc.c
@@ -1,3 +1,0 @@
-void func_ov006_020dcffc(void)
-{
-}
